partially translated documents
used Google for the moment , My spanish is Good to about an midlevel
with a few gaps.
my Italian or Japanese is a bit rusted  私はさびた日本語を話す。
QA pre-checks on Spanish looks good.  Italian seems good but with
Dialects  issues in regions , its worth a good hard look.
the NIHONGO /Japanese same bit.